% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colar.R
\name{colar}
\alias{colar}
\title{Colar dois parametros}
\usage{
colar(a, b)
}
\arguments{
\item{a}{primeiro parametro}

\item{b}{segundo parametro}
}
\value{
descricao do resultado
}
\description{
Esta funcao recebe dois parametros e cola
}
