% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soma.R
\name{soma}
\alias{soma}
\title{Soma dois numeros}
\usage{
soma(x, y)
}
\arguments{
\item{x}{primeiro argumento}

\item{y}{segundo argumento}
}
\value{
a soma de dois numeros
}
\description{
recebe um numero x e outro y e retorna a soma de x e y
}
